Connect when data is ready
We want to make the conneciton when the data is in. First when opening
the file we see the PAR1 Parquet header, and then a few seconds later
the schema of the parquet file is written. Most of the time it takes
a while before pulling the blocks from HDFS and serve them the the
SWIFT client. Most of the time, this takes too long..